+++ dominio/src/main/dominio/Entidad.java
@@ -1,15 +1,15 @@
 package main.dominio;
 
 public abstract class Entidad {
-	private int posicionX;
-	private int posicionY;
+	private double posicionX;
+	private double posicionY;
 
-	public Entidad(int posX, int posY) {
+	public Entidad(double posX, double posY) {
 		this.posicionX = posX;
 		this.posicionY = posY;
 	}
 
-	public int getPosicionX() {
+	public double getPosicionX() {
 		return posicionX;
 	}
 
@@ -17,7 +17,7 @@ public abstract class Entidad {
 		this.posicionX = posicionX;
 	}
 
-	public int getPosicionY() {
+	public double getPosicionY() {
 		return posicionY;
 	}
